RMS server

+++ routes/Addagent.js
@@ -1,5 +1,6 @@
 var express = require("express");
 var router = express.Router();
+var Tenants = require("../modals/Tenants");
 var Addagent = require("../modals/Addagent");
 var {
   verifyToken,
@@ -53,9 +54,41 @@ router.post("/addagent", async (req, res) => {
 // delete Agent 
 router.delete("/delete_agent", async (req, res) => {
     try {
-      let result = await Addagent.deleteMany({
-        _id: { $in: req.body },
+
+      const propIdsToDelete = req.body;
+
+      console.log("propIdsToDelete from body", propIdsToDelete);
+  
+      // Get the names of the staff members to be deleted
+      const propertyToDelete = await Addagent.find({
+        _id: { $in: propIdsToDelete },
+      }).select("agent_name");
+      console.log("propertyToDelete after variable",propertyToDelete)
+  
+      const propNamesToDelete = propertyToDelete.map(
+        (staff) => staff.agent_name
+      );
+  
+      const assignedProperty = await Tenants.find({
+        leasing_agent: { $in: propNamesToDelete },
+      }); 
+  
+      if (assignedProperty.length > 0) {
+        return res.status(201).json({
+          statusCode: 201,
+          message:
+            "Agent are already assigned. Deletion not allowed.",
+        });
+      }
+
+      const result = await Addagent.deleteMany({
+        _id: { $in: propIdsToDelete },
       });
+
+      // let result = await Addagent.deleteMany({
+      //   _id: { $in: req.body },
+      // });
+
       res.json({
         statusCode: 200,
         data: result,
